requirements.txt
python-nmap==0.7.1
# python3-nmap==1.5.1
PyYAML==5.4
requests==2.27.1

